[REFACTOR] Move downloaders to separate files
