+++ src/ytdl_sub/config/preset_class_mappings.py
@@ -3,12 +3,12 @@ from typing import List
 from typing import Type
 
 from ytdl_sub.downloaders.downloader import Downloader
-from ytdl_sub.downloaders.soundcloud_downloader import SoundcloudAlbumsAndSinglesDownloader
+from ytdl_sub.downloaders.soundcloud.albums_and_singles import SoundcloudAlbumsAndSinglesDownloader
+from ytdl_sub.downloaders.youtube.channel import YoutubeChannelDownloader
 from ytdl_sub.downloaders.youtube.merge_playlist import YoutubeMergePlaylistDownloader
+from ytdl_sub.downloaders.youtube.playlist import YoutubePlaylistDownloader
 from ytdl_sub.downloaders.youtube.split_video import YoutubeSplitVideoDownloader
-from ytdl_sub.downloaders.youtube_downloader import YoutubeChannelDownloader
-from ytdl_sub.downloaders.youtube_downloader import YoutubePlaylistDownloader
-from ytdl_sub.downloaders.youtube_downloader import YoutubeVideoDownloader
+from ytdl_sub.downloaders.youtube.video import YoutubeVideoDownloader
 from ytdl_sub.plugins.music_tags import MusicTagsPlugin
 from ytdl_sub.plugins.nfo_tags import NfoTagsPlugin
 from ytdl_sub.plugins.output_directory_nfo_tags import OutputDirectoryNfoTagsPlugin
