+++ src/ytdl_sub/downloaders/youtube/split_video.py
@@ -4,8 +4,8 @@ from pathlib import Path
 from typing import Dict
 from typing import List
 
-from ytdl_sub.downloaders.youtube_downloader import YoutubeDownloader
-from ytdl_sub.downloaders.youtube_downloader import YoutubeVideoDownloaderOptions
+from ytdl_sub.downloaders.youtube.abc import YoutubeDownloader
+from ytdl_sub.downloaders.youtube.video import YoutubeVideoDownloaderOptions
 from ytdl_sub.entries.youtube import YoutubePlaylistVideo
 from ytdl_sub.entries.youtube import YoutubeVideo
 from ytdl_sub.utils.chapters import Chapters
